♻️ refactor: アプリケーション層配下にユースケース層を持たせるようにした
#11
