+++ lib/application/usecase/post/state/posts.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../domain/post/post_repository.dart';
+import '../../../../domain/post/post_repository.dart';
 
 /// PostRepositoryから投稿一覧を取得するプロバイダー
 final postsProvider = StreamProvider(
