+++ lib/application/usecase/run_usecase_mixin.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_reference_app/usecase/state/overlay_loading_provider.dart';
+import 'package:flutter_reference_app/application/state/overlay_loading_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 /// ユースケース実行のためのメソッドを備えた Mixin
