+++ lib/presentation/page/post/component/post_text_form_field.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../usecase/state/selected_post.dart';
+import '../../../../application/usecase/post/state/selected_post.dart';
 
 final contentControllerProvider =
     StateProvider.autoDispose<TextEditingController>(
