+++ lib/presentation/page/time_line/component/post_list_view.dart
@@ -8,8 +8,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hexcolor/hexcolor.dart';
 
-import '../../../../usecase/state/posts.dart';
-import '../../../../usecase/state/selected_post.dart';
+import '../../../../application/usecase/post/state/posts.dart';
+import '../../../../application/usecase/post/state/selected_post.dart';
 
 /// 投稿一覧リスト
 class PostListView extends ConsumerWidget with ErrorHandlerMixin {
